+cc solma, ckuhn, raz
I'm not sure if Foo is OOO starting Monday or Tuesday. Just in case, I've added you to reviewers. I'll delay submitting this until at least one of you four gives the thumbs up. I won't be in on Monday, but I will check when I can so that we can resubmit as early as possible to avoid causing merge conflicts with outstanding CLs that sync'd past this.

Note there is a two line change to the rollback below to not update the trainer just yet. All tests past.

Automated g4 rollback of changelist 223605799.

*** Reason for rollback ***

Should have thought of this on Friday, but if EvalInputReceiver is temporarily updated to call the LegacyEvalInputReceiver then this CL can still be used to push the forward compatible evaluator changes.

See update at line 115 in third_party/py/tensorflow_model_analysis/eval_saved_model/export.py

*** Original change description ***

Automated g4 rollback of changelist 223456524.

*** Reason for rollback ***

Not compatible with TFX pipelines.

*** Original change description ***

TFMA - EvalInputReceiver updates
  * Move features/labels to SignatureDef instead of collections dict

***

***

PiperOrigin-RevId: 223841315