Use static shape to determine if we need to convert label to one-hot vector.
When using tf.cond, the graphs corresponding to both conditions are created, and the decision of which branch to take is done dynamically on a per-batch basis. This means that even if no one-hot conversion is necessary, that graph is still made and will produce errors if the input is not appropriate. For example, if the labels tensor is a bool tensor of shape [batch_size, num_classes], the graph creation of the one-hot vector will fail due to an unexpected bool type, even though that graph is never run.

Note that we're now relying on the 2nd dim being unknown in order to indicate that the conversion should happen. W/ that assumption we can now support label tensors w/ different numbers of classes per example, hence the change to "multi-hot" instead of "one-hot".

I added an assertion to make sure the dynamic shapes of the predictions and labels tensors are the same.

PiperOrigin-RevId: 226177749