Support for Bounded Value to represent float. The bounds may be absent if the user did not enable confidence interval computation.
Introduce a helper, tfma.CellRenderer.extractFloatValue, that can be used both in tfma support library and in components.

Note that The code assumes that for field xyz_metric, the bounded value version will be named bounded_xyz_metric in proto which will become boundedXyzMetric when converted to JSON.

The relavant proto changes were introduced in cl/226359455 and cl/227874831 fills the bounded version.

PiperOrigin-RevId: 230407882